Se implentaron componentes
