+++ app/index.js
@@ -1,75 +1,5 @@
-import React, { useState } from 'react';
-import { View, Text, TextInput, Button, StyleSheet } from 'react-native';
-import { useRouter } from 'expo-router';
+import { LogIn } from "./components/LogIn"
 
-export default function Login() {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState('');
-  const router = useRouter(); // Para redirigir al Home después del inicio de sesión
-
-  const handleLogin = () => {
-    // Validación simple de inicio de sesión
-    if (username === 'Admin' && password === 'Admin') {
-      // Redirige al Home después del inicio de sesión exitoso
-      router.push('/subs');
-    } else {
-      setError('Usuario o contraseña incorrectos');
-    }
-  };
-
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Iniciar Sesión</Text>
-      
-      {/* Campo de usuario */}
-      <TextInput
-        style={styles.input}
-        placeholder="Usuario"
-        value={username}
-        onChangeText={(text) => setUsername(text)}
-      />
-
-      {/* Campo de contraseña */}
-      <TextInput
-        style={styles.input}
-        placeholder="Contraseña"
-        secureTextEntry
-        value={password}
-        onChangeText={(text) => setPassword(text)}
-      />
-
-      {/* Botón de inicio de sesión */}
-      <Button title="Iniciar Sesión" onPress={handleLogin} />
-
-      {/* Mensaje de error */}
-      {error ? <Text style={styles.error}>{error}</Text> : null}
-    </View>
-  );
+export default function newSub() {
+  return <LogIn />
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    padding: 20,
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 20,
-    textAlign: 'center',
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: '#ccc',
-    borderRadius: 5,
-    padding: 10,
-    marginBottom: 20,
-  },
-  error: {
-    marginTop: 20,
-    textAlign: 'center',
-    color: 'red',
-  },
-});
